Some more speedups. Moving to floats.

+++ lib/local/CLM/include/CCNF_patch_expert.h
@@ -97,7 +97,7 @@ public:
 
 	void Read(std::ifstream &stream);
 	// The im_dft, integral_img, and integral_img_sq are precomputed images for convolution speedups (they get set if passed in empty values)
-	void Response(Mat_<float> &im, Mat_<double> &im_dft, Mat &integral_img, Mat &integral_img_sq, Mat_<double> &resp);
+	void Response(Mat_<float> &im, Mat_<double> &im_dft, Mat &integral_img, Mat &integral_img_sq, Mat_<float> &resp);
 
 };
 
@@ -146,7 +146,7 @@ public:
 	void Read(std::ifstream &stream, std::vector<int> window_sizes, std::vector<std::vector<Mat_<float> > > sigma_components);
 
 	// actual work (can pass in an image and a potential depth image, if the CCNF is trained with depth)
-	void Response(Mat_<float> &area_of_interest, Mat_<double> &response);    
+	void Response(Mat_<float> &area_of_interest, Mat_<float> &response);    
 
 	// Helper function to compute relevant sigmas
 	void ComputeSigmas(std::vector<Mat_<float> > sigma_components, int window_size);
